+++ headerFile/Orc.h
@@ -7,6 +7,6 @@ class Orc : public Monster
 {
 
 public:
-	Orc() {}// : Monster("Orc") { name += " 오크"; }
+	Orc() : Monster("Orc") { name += " 오크"; }
 	Orc(const std::string& _name) : Monster(_name) {}
 };
